refactor: rename share handler file to share-playlist for clarity
