+++ src/apps/gateway/videos/index.ts
@@ -8,7 +8,7 @@ import { HasuraWebhookRequest, hasuraWebhookSchema } from 'src/schema/hasura';
 import { ConvertRequest, convertSchema } from 'src/schema/videos/convert';
 import { CrawlRequest, crawlSchema } from 'src/schema/videos/crawl';
 import { ShareRequest, shareSchema } from 'src/schema/videos/share';
-import { sharePlaylistHandler } from './routes/share';
+import { sharePlaylistHandler } from './routes/share-playlist';
 
 const videosRouter: Router = express.Router();
